requirements.txt
flask
flask-cors
python-dotenv
faster-whisper
requests
av
numpy
scikit-learn
python_speech_features
dateparser
pydantic